tests/wpt/tests/css/css-viewport/computedStyle-zoom.html

<!DOCTYPE html>
<title>getComputedStyle for elements with css zoom</title>
<link rel="author" title="Yotam Hacohen" href="mailto:yotha@chromium.org">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://drafts.csswg.org/css-viewport/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<head>
  <style>
    div {
    width: 64px;
    height: 64px;
    background-color: blue
    }
    div.x4_zoom {
    zoom: 4.0;
    background-color: blueviolet;
    }
    div.x2_zoom {
    background-color: chartreuse;
    zoom: 2.0;
    }

  </style>
</head>
<body>
  <div id="no_zoom"></div>
  <div class="x4_zoom" id="with_zoom"></div>
  <div class="x2_zoom" id="parent_div">
    <div class="x4_zoom" id="nested_zoom"></div>
  </div>
  <div class="x2_zoom" id="testing_set_style" style="height: 1px; width: 1px;"></div>
  <script>
    test(function() {
      assert_true(!!no_zoom, "no zoom target exists");
      assert_true(!!with_zoom, "zoom target exists");
      assert_true(!!nested_zoom, "zoom target exists");
      assert_true(!!parent_div, "parent div with zoom exists")
    });
    test(function(){
      noZoomStyle = window.getComputedStyle(no_zoom);
      assert_equals(noZoomStyle.getPropertyValue("width"), "64px");
      assert_equals(noZoomStyle.getPropertyValue("height"), "64px");
      assert_equals(noZoomStyle.getPropertyValue("zoom"), "1");
    });
    test(function(){
      withZoomStyle = window.getComputedStyle(with_zoom);
      assert_equals(withZoomStyle.getPropertyValue("width"), "64px");
      assert_equals(withZoomStyle.getPropertyValue("height"), "64px");
      assert_equals(withZoomStyle.getPropertyValue("zoom"), "4");
    });
    test(function(){
      parentWithZoomStyle = window.getComputedStyle(parent_div);
      assert_equals(parentWithZoomStyle.getPropertyValue("width"), "64px");
      assert_equals(parentWithZoomStyle.getPropertyValue("height"), "64px");
      assert_equals(parentWithZoomStyle.getPropertyValue("zoom"), "2");
    });
    test(function(){
      nestedZoomStyle = window.getComputedStyle(nested_zoom);
      assert_equals(nestedZoomStyle.getPropertyValue("width"), "64px");
      assert_equals(nestedZoomStyle.getPropertyValue("height"), "64px");
      assert_equals(nestedZoomStyle.getPropertyValue("zoom"), "4");
    });
    test(function(){
      testDivStyle = window.getComputedStyle(testing_set_style);
      assert_equals(testDivStyle.getPropertyValue("width"), "1px");
      assert_equals(testDivStyle.getPropertyValue("height"), "1px");
      assert_equals(testDivStyle.getPropertyValue("zoom"), "2");
      window.testing_set_style.setAttribute("style", "width: 64px; height: 64px;");
      assert_equals(testDivStyle.getPropertyValue("width"), "64px");
      assert_equals(testDivStyle.getPropertyValue("height"), "64px");
      assert_equals(testDivStyle.getPropertyValue("zoom"), "2");
    });
  </script>
</body>
